+++ CHANGELOG.md
@@ -1,8 +1,12 @@
 # Changelog
 # 3.3.5
 - zsim:
-  - Fixed instruction time display of HALT
+  - Fixed instruction time display of HALT.
   - Fixed FLASH frequency from 1/625ms to 1/640ms.
+  - Simulation of 'zxnDMA'.
+- Logging:
+  - Settings: 'dezog.log.customCode' changed to 'dezog.log.zsim.customCode'.
+  - Added new log for the zxnDMA under 'dezog.log.zsim.hardware'.
 
 # 3.3.4
 - Fix for #126: Data misalignment for ZX128K SNA with Active BANK6 in zsim. Banks in SNA loading fixed.